Enable coinjoin on regtest env (#6)
* fix: prevent 'config value not set' on cj in regtest env

'max_cj_fee' config values must be present when invoking a
coinjoin via the api. the default config for the regtest env
did not include these values before, hence they could not
be overridden in docker-compose setup file.

* dev: switch irc server from miniirc to ergochat

* dev: start second joinmarket container

In order to successfully perform a CoinJoin, there must be at least one
maker. This commit adds a second JoinMarket container with
jmwalletd accessible on port 29183.

* dev: add helper script to mine a block

* dev: add container switch to regtest-control script

* dev: invoke mine-block in regtest-control script

* dev: add flag '--unmatured' to regtest-control script

Sometimes it is desired that block rewards do not reach maturity
immediately. This is convenient if you want to supply several wallets
with coins, but do not want to generate many blocks all the time.

* dev: add helper script to start maker service

This script helps in providing both JoinMarket containers a wallet with
spendable coins and starting the Maker Service in the secondary
container. Its main goal is to make CoinJoin transactions possible in the
regtest environment.

* dev: move common functions to script common.sh

* fix: use correct irc host in docker setup

* dev: use ergochat image from ghcr.io

* dev: dont lock secondary wallet in setup script

normally the wallet should be locked, but it seems this leads to the
maker service responding very slowly which causes failing coinjoins.
as the secondary instance is never interacted with directly, we can
refrain from unlocking the wallet for now.

* docs: add debug log command to docker readme

* cleanup: rename script 'regtest-control' to 'fund-wallet'

* dev: upgrade joinmarket from v0.9.3 to use current master branch

* review: add more descriptive comments and cleanup files