feat: readd websocket connection (#122)
* feat: readd websocket connection

* add WebsocketProvider

* feat: update taker/maker state based on websocket data

* update connection indicator based on websocket state

* reconnect with linear backoff on websocket errors

the backend is not happy with a second instance in another browser tab
that send invalid tokens (e.g. because the previously used wallet was
locked in another tab). This introduces a linear increase to a max
of 10 seconds. Seems to ease the problem in the active tab.

* send keepalive message periodically

some webservers will close connections on inactivity rather
soon by default (e.g. nginx is 60s). this commit contains
sending a keepalive message periodically every 30 seconds.

Co-authored-by: d11n <mail@dennisreimann.de>