dev(regtest): add npm script regtest:build
this command will build the development setup with docker cache
enabled. usually must fast as it will only pull in newer base
images and not download and recompile jm again. this is the
preferred way if you made only a few adaptions to the images
and e.g. not want to fetch the current master of jm.