fix: create non-descriptor wallet (#487)
* dev(regtest): remove nbxplorer in favor of simpler custom regtest-initializer

* dev(regtest): use jsonrpc 2.0 requests to create and load wallet

* dev(regtest): upgrade bitcoin core from v22 to v23

* fix(regtest): wait-for-bitcoind should wait for at least >=0 blocks

* dev(regtest): use jam-dev-standalone

* dev(regtest): add npm script regtest:build

this command will build the development setup with docker cache
enabled. usually must fast as it will only pull in newer base
images and not download and recompile jm again. this is the
preferred way if you made only a few adaptions to the images
and e.g. not want to fetch the current master of jm.